to_int check, used std::strtol for long and errors.

+++ ex00/ScalarConverter.cpp
@@ -6,7 +6,7 @@
 /*   By: macbookpro <macbookpro@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/29 01:31:28 by ael-moha          #+#    #+#             */
-/*   Updated: 2025/05/30 21:54:14 by macbookpro       ###   ########.fr       */
+/*   Updated: 2025/05/30 22:03:11 by macbookpro       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -58,7 +58,7 @@ void to_char(const std::string &literal)
 
 void to_int (const std::string &literal)
 {
-    long d = static_cast<int>(std::atoi(literal.c_str()));
+    long d = static_cast<long>(std::atol(literal.c_str()));
 
     if (d < 0 || d > 127)
         std::cout << "char: impossible" << std::endl;
